refactor(site/src/pages/AgentsPage): rename display mode settings module
