+++ site/src/pages/AgentsPage/components/DisplayModeSettings.tsx
